implement switch account functionality

+++ packages/web/src/components/SettingsUsers.tsx
@@ -83,8 +83,11 @@ function SettingsUsers() {
   }
 
   const deleteUser = async () => {
+    if (!selectedUser) return
     if (selectedUser?.role === 'notifications') {
       await axios.delete(`/settings/notifications/emails/${selectedUser.id}`)
+    } else {
+      await axios.delete(`/settings/users/${selectedUser.id}`)
     }
     toast({
       position: 'top',
